requirements.txt.txt
Django
djangorestframework
django-filter
django-crispy-forms
django-allauth
crispy_bootstrap4